+++ src/app/datalogger/ina219.h
@@ -22,4 +22,4 @@
 #define VOLT_PRESCALE_DIV8 (0b11 << 11)
 
 bool ina219_init(uint8_t device_addr, uint32_t prescale, uint16_t calibration);
-bool ina219_read_microamps(uint8_t device_addr, int32_t *val /* OUT */);
+bool ina219_read_microamps(uint8_t device_addr, int16_t *val /* OUT */);
